+++ Assets/uNode3/changelog.md
@@ -2,6 +2,7 @@
 
 ## v3.0.7
 - Added default value support for parameters
+- Added Graph Inspector window, useful if needed to see both node properties and game object properties on the screen in the same time
 - Improved c# parser
 
 ## v3.0.6